refactor(ipc): split mod.rs (1288 LOC) — extract protocol + diagnostics to sibling files
ipc/ was a single-file module dir — mod.rs at 1288 LOC mixing wire
protocol, memory diagnostics, server state, connection handler, tests,
and server lifecycle in one file. Flagged in claude-tab-1's broad-code
audit broadcast 2026-05-18 19:40Z ("ipc/mod.rs 1288 LOC AND it is the
ONLY file in ipc/").

Per Joel's zero-users no-migration-ceremony directive, ONE PR cuts
the size + re-exports the moved types so call sites resolve unchanged.

## What changes

- New `ipc/protocol.rs` — `InboxMessageRequest` (ts-rs) + internal
  `Response` struct. Moves both away from mod.rs.
- New `ipc/diagnostics.rs` — per-command RSS tracking (`current_rss_mb`,
  `log_command_rss_delta`, `dump_memory_report`). Pure observability,
  no wire impact.
- `ipc/mod.rs` shrinks from 1288 → ~1135 LOC; declares the two new
  submodules + re-exports `InboxMessageRequest` so existing
  `crate::ipc::InboxMessageRequest` resolves unchanged.
- Removed now-unused `serde::{Deserialize, Serialize}` + `ts_rs::TS`
  imports from mod.rs (their types moved out).

## What stays in mod.rs (deliberately)

- `ServerState` struct + all module Arc references — too tightly
  coupled to `start_server` to extract in one PR. Future split would
  carve out `state.rs` separately.
- Connection handler (binary framing + dispatch loop) — needs the
  module Arc references in scope.
- `start_server` lifecycle.
- Inline `#[cfg(test)]` tests for binary framing — convention to keep
  inline with the code they test.

## Discipline

- Zero behavior change. Pure file-shape refactor.
- All visibility preserved: `pub use protocol::InboxMessageRequest`
  in mod.rs ensures external callers (TS bindings, generated barrel)
  see the type at the same path.
- `Response` stays `pub(crate)` (was private in mod.rs; module
  boundary requires `pub(crate)` so connection handler can use it).
- Clippy held at 157 baseline.

## Verification

- `cargo check --lib --features metal,accelerate` — clean
- `cargo test --lib --features metal,accelerate ipc::` — 9/9 pass
  (binary frame roundtrip + ts-rs export bindings)
- `cargo clippy --lib --features metal,accelerate` — 157 warnings
  (baseline held)

## Refs

- claude-tab-1 broad-code-audit broadcast 2026-05-18 19:40Z (named
  ipc/mod.rs as one of the god-modules)
- Joel 2026-05-18 19:44Z (zero-users full-blown-Rust-dev mode →
  one PR per refactor, no migration cadence)
